Added packages for security,userdao,userservice, klp
PROSOXI STO UserServiceImpl sti methodo update, exw sxoliasei kapoia pragmata pou xtipousan. Aforoun sta pedia pou exmeis exoume ston pinaka customers. Logika de tha xreiastei na kanoume kapoia prosthiki ekei, mias kai otan kapoios allazei username i password de xreiazetai na alaksei kati ston pinaka customers.

+++ src/main/java/com/mycompany/bioeshop/entities/User.java
@@ -47,7 +47,7 @@ public class User implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(nullable = false)
-    private Long id;
+    private Integer id;
     @Basic(optional = false)
     @NotNull
     @Size(min = 1, max = 30)
@@ -70,21 +70,21 @@ public class User implements Serializable {
     public User() {
     }
 
-    public User(Long id) {
+    public User(int id) {
         this.id = id;
     }
 
-    public User(Long id, String ssoId, String password) {
+    public User(int id, String ssoId, String password) {
         this.id = id;
         this.ssoId = ssoId;
         this.password = password;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
